Bamboo SendinBlue Adapter v0.2.0 Bamboo.SendinBlueAdapter

Sends email using SendinBlue’s JSON API v2.0.

Use this adapter to send emails through SendinBlue’s API. Requires that an API key is set in the config.

If you would like to add a replyto header to your email, then simply pass it in using the header property or put_header function like so:

put_header("reply-to", "foo@bar.com")

Example config

# In config/config.exs, or config.prod.exs, etc.
config :my_app, MyApp.Mailer,
  adapter: Bamboo.SendinBlueAdapter,
  api_key: "my_api_key"

# Define a Mailer. Maybe in lib/my_app/mailer.ex
defmodule MyApp.Mailer do
  use Bamboo.Mailer, otp_app: :my_app
end

Functions

deliver(email, config)

Callback implementation for Bamboo.Adapter.deliver/2.